[qtbase] Don't use QStringLiteral in dynamically unloadable code
QStringLiteral generates a QString::Data block with -1 ref count which
lives somewhere in the data section of the libqconnmanbearer.so and becomes
invalid as soon as plugin gets unloaded. However, pointers to these static
QString::Data blocks get carried along with QDBusPendingCall and may outlive
the data section they point to. That results in a crash, typically on exit.