[spec] Fix scriptlets in qt5-qtconcurrent and others (Fixes JB#27995)
From https://en.opensuse.org/openSUSE:Packaging_scriptlet_snippets:

Have in mind that sections in a RPM spec file end where the next section
begins. This can, in some cases, lead to unexpected errors. For example, if you
have comments [...]

In this case, the "# The postun section:" comment, for RPM, actually belongs to
the %post section (and the "# The post section:" comment actually belongs to
the section before %post). In some cases, e.g. in the build system for SLE11
and SLE11_SP1, such a comment results in a situtation where an unintended
argument '0' or '1' is added to the /sbin/ldconfig call which let it fail with
errors like "/sbin/ldconfig: relative path '1' used to build cache". [...]